TimeCode fetch from database Ok

+++ AppMovil/composeApp/src/commonMain/kotlin/com/es/appmovil/viewmodel/ResumeViewmodel.kt
@@ -38,8 +38,8 @@ class ResumeViewmodel {
         return days
     }
 
-    fun getLegend(): MutableState<MutableMap<String, Long>> {
-        val legend = mutableStateOf(mutableMapOf<String, Long>())
+    fun getLegend(): MutableState<MutableMap<String, ULong>> {
+        val legend = mutableStateOf(mutableMapOf<String, ULong>())
 
         timeCodes.value.forEach {
             legend.value[it.desc] = it.color
@@ -55,11 +55,11 @@ class ResumeViewmodel {
             .forEach {
                 val timeCode = timeCodes.value.find { time -> time.idTimeCode == it.idTimeCode }
                 if (timeCode != null) {
-                    if (timeActivity.value.containsKey(timeCode.color)) {
-                        val hours = timeActivity.value[timeCode.color]?.plus(it.time)
-                        timeActivity.value[timeCode.color] = hours ?: 0f
+                    if (timeActivity.value.containsKey(timeCode.color.toLong())) {
+                        val hours = timeActivity.value[timeCode.color.toLong()]?.plus(it.time)
+                        timeActivity.value[timeCode.color.toLong()] = hours ?: 0f
                     } else {
-                        timeActivity.value[timeCode.color] = it.time
+                        timeActivity.value[timeCode.color.toLong()] = it.time
                     }
                 }
         }
@@ -75,7 +75,7 @@ class ResumeViewmodel {
                 val timeCode = timeCodes.value.find { time -> time.idTimeCode == it.idTimeCode }
                 timeCode?.let { tc ->
                     val currentList = dayActivity.value[it.date] ?: mutableListOf()
-                    currentList.add(Color(tc.color))
+                    currentList.add(Color(tc.color.toLong()))
                     dayActivity.value[it.date] = currentList
                 }
             }
